+++ source/classes/Petition.js
@@ -1,0 +1,11 @@
+class Petition {
+	/** @param {string} nameString */
+	constructor(nameString) {
+		this.name = nameString;
+	}
+	petitionerIds = [];
+}
+
+module.exports = {
+	Petition
+};