+++ source/commands/club-skip-meeting.js
@@ -1,7 +1,7 @@
 const { CommandWrapper } = require('../classes/index.js');
 const { cancelClubEvent, createClubEvent, clearClubReminder, scheduleClubReminderAndEvent, updateClubDetails } = require('../engines/clubEngine.js');
 const { isClubHostOrModerator } = require('../engines/permissionEngine.js');
-const { getClubDictionary, updateClub, updateList } = require('../engines/referenceEngine.js');
+const { getClubDictionary, updateClub, updateListReference } = require('../engines/referenceEngine.js');
 const { timeConversion } = require('../util/mathUtil.js');
 
 const mainId = "club-skip-meeting";
@@ -24,7 +24,7 @@ module.exports = new CommandWrapper(mainId, "Skip the next club meeting, cancell
 			club.timeslot.nextMeeting = null;
 		}
 		updateClubDetails(club, interaction.channel);
-		updateList(interaction.guild.channels, "club");
+		updateListReference(interaction.guild.channels, "club");
 		updateClub(club);
 		interaction.reply({ content: `This club's next meeting will be skipped.${club.timeslot.periodCount && club.timeslot.periodUnits ? `The next meeting will instead be <t:${club.timeslot.nextMeeting}>.` : ""}` });
 	}