+++ source/engines/referenceEngine.js
@@ -1,121 +1,10 @@
-const { GuildChannelManager, Guild, User, Collection, ChannelType, ActionRowBuilder, StringSelectMenuBuilder, Message, MessageFlags } = require('discord.js');
+const { GuildChannelManager, ActionRowBuilder, StringSelectMenuBuilder, Message, MessageFlags } = require('discord.js');
 const { Club, ClubTimeslot } = require("../classes");
-const { topicCategoryId } = require('../constants');
-const { ensuredPathSave } = require('../util/fileUtil.js');
+const { MAX_EMBEDS_PER_MESSAGE, MAX_EMBED_FIELD_NAME_LENGTH, MAX_EMBED_FIELD_VALUE_LENGTH, MAX_EMBED_TOTAL_CHARACTERS, MAX_EMBED_DESCRIPTION_LENGTH } = require('../constants');
 const { embedTemplateBuilder } = require("./messageEngine.js");
+const { getRolePetitions, getChannelPetitions } = require('./customizationEngine.js');
+const { ensuredPathSave } = require('../util/fileUtil.js');
 const { commandMention } = require('../util/textUtil.js');
-
-/**  key: topic, value: petitioner ids
- * @type {Record<string, string[]>} */
-let petitions = require('../../config/petitionList.json');
-
-function getPetitions() {
-	return petitions;
-}
-
-/** Add a petition to the petition list and update the topic list embed
- * @param {string} petitionListInput
- * @param {GuildChannelManager} channelManager
- */
-function setPetitions(petitionListInput, channelManager) {
-	petitions = petitionListInput;
-	ensuredPathSave(petitions, 'petitionList.json');
-	updateList(channelManager, "petition");
-}
-
-/** Create a topic channel for a petition if it has enough ids
- * @param {Guild} guild
- * @param {string} topicName
- * @param {User} author
- * @returns {{petitions: number, threshold: number}}
- */
-function checkPetition(guild, topicName, author = null) {
-	if (!petitions[topicName]) {
-		petitions[topicName] = [];
-	}
-	if (author) {
-		if (!petitions[topicName].includes(author.id)) {
-			petitions[topicName].push(author.id);
-		} else {
-			author.send(`You have already petitioned for ${topicName}.`)
-				.catch(console.error)
-			return;
-		}
-	}
-	const petitionCount = petitions[topicName].length ?? 0;
-	const threshold = Math.ceil(guild.memberCount * 0.05) + 1;
-	if (petitionCount >= threshold) {
-		addTopicChannel(guild, topicName);
-	} else {
-		setPetitions(petitions, guild.channels);
-	}
-	updateList(guild.channels, "petition");
-	return {
-		petitions: petitionCount,
-		threshold
-	}
-}
-
-// Collection <channelId, channelName>
-const topics = new Collection();
-
-/** Get the array of topic channel ids
- * @returns {string[]}
- */
-function getTopicIds() {
-	return Array.from(topics.keys());
-}
-
-/** Get the array of topic channel names
- * @returns {string[]}
- */
-function getTopicNames() {
-	return Array.from(topics.values());
-}
-
-/** Add a new entry to the topic map
- * @param {string} id
- * @param {string} channelName
- */
-function addTopic(id, channelName) {
-	topics.set(id, channelName);
-}
-
-/** Clean up internal state to keep in sync with removing a topic channel
- * @param {string} channelId
- * @param {Guild} guild
- */
-function removeTopic(channelId, guild) {
-	topics.delete(channelId);
-	ensuredPathSave(topics, 'topicList.json');
-	updateList(guild.channels, "petition");
-}
-
-/** Add the new topic channel topic list to prevent duplicate petitions
- * @param {Guild} guild
- * @param {string} topicName
- * @returns {Promise<TextChannel>}
- */
-function addTopicChannel(guild, topicName) {
-	return guild.channels.create({
-		name: topicName,
-		parent: topicCategoryId,
-		type: ChannelType.GuildText
-	}).then(channel => {
-		if (!petitions[topicName]) {
-			petitions[topicName] = [];
-		}
-
-		if (petitions[topicName].length > 0) {
-			channel.send(`This channel has been created thanks to: <@${petitions[topicName].join('> <@')}>`);
-		}
-		delete petitions[topicName];
-		addTopic(channel.id, channel.name);
-		ensuredPathSave(topics, 'topicList.json');
-		setPetitions(petitions, guild.channels);
-		return channel;
-	}).catch(console.error);
-}
 
 /** @type {{[clubId: string]: Club}>} */
 const clubDictionary = {};
@@ -143,80 +32,143 @@ function updateClub(club) {
 function removeClub(id, channelManager) {
 	delete clubDictionary[id];
 	ensuredPathSave(clubDictionary, 'clubList.json');
-	updateList(channelManager, "club");
+	updateListReference(channelManager, "club");
 }
 
 /** @type {{petition: {channelId: string; messageId: string}, club: {channelId: string; messageId: string;}, rules: {channelId: string; messageId: string;}, "press-kit": {channelId: string; messageId: string;}, "proxy-thread-info": {channelId: string; messageId: string;}}} */
 let referenceMessages = require('../../config/referenceMessageIds.json');
 
-/** Builds the MessageOptions for the specified list message
+/** Builds the MessageOptions for the petition list message
  * @param {number} memberCount
- * @param {"petition" | "club"} listType
  * @returns {Promise<import('discord.js').BaseMessageOptions>}
  */
-function buildListMessagePayload(memberCount, listType) {
-	let description;
+function buildPetitionListPayload(memberCount) {
+	const messageOptions = { flags: MessageFlags.SuppressNotifications };
 
-	const messageOptions = {
-		flags: MessageFlags.SuppressNotifications,
-		components: [new ActionRowBuilder()]
-	}
-
-	const selectMenu = new StringSelectMenuBuilder().setCustomId(`${listType}List`)
+	const channelSelect = new StringSelectMenuBuilder().setCustomId("petitionChannel")
+		.setMinValues(1);
+	const roleSelect = new StringSelectMenuBuilder().setCustomId("petitionRole")
 		.setMinValues(1);
 
-	switch (listType) {
-		case "petition":
-			description = `Here are the topic channels that have been petitioned for. Note that petitions will be converted to lowercase to match with Discord text channels being all lowercase. They will automatically be added when reaching **${Math.ceil(memberCount * 0.05)} petitions** (5% of the server). You can sign onto an already open petition with the select menu under this message (jump to message in pins).\n`;
+	const fields = [{ name: "Channel Petitions", value: "" }, { name: "Role Petitions", value: "" }];
+	for (const petition of getChannelPetitions()) {
+		fields[0].value += `\n${petition.name}: ${petition.petitionerIds.length} petitioner${petition.petitionerIds.length === 1 ? "" : "s"} so far`;
+		channelSelect.addOptions([{
+			label: petition.name,
+			value: petition.name
+		}]);
+	}
+	if (fields[0].value === "") {
+		fields[0].value = "No open channel petitions yet";
+	}
+	for (const petition of getRolePetitions()) {
+		fields[1].value += `\n${petition.name}: ${petition.petitionerIds.length} petitioner${petition.petitionerIds.length === 1 ? "" : "s"} so far`;
+		roleSelect.addOptions([{
+			label: petition.name,
+			value: petition.name
+		}]);
+	}
+	if (fields[1].value === "") {
+		fields[1].value = "No open role petitions yet";
+	}
 
-			for (const topicName in petitions) {
-				description += `\n${topicName}: ${petitions[topicName].length} petitioner(s) so far`;
-				selectMenu.addOptions([{
-					label: topicName,
-					value: topicName
-				}]);
-			}
+	if (channelSelect.options.length > 0) {
+		channelSelect.setPlaceholder("Select a channel petition...");
+	} else {
+		channelSelect.setPlaceholder("No open channel petitions");
+	}
 
-			if (selectMenu.options.length > 0) {
-				selectMenu.setPlaceholder("Select a petition...");
-			} else {
-				selectMenu.setPlaceholder("No open petitions");
-			}
-			break;
-		case "club":
-			description = "Here's a list of the clubs on the server. Learn more about one by typing `/club-invite (club ID)`.\n";
+	if (roleSelect.options.length > 0) {
+		roleSelect.setPlaceholder("Select a role petition...");
+	} else {
+		roleSelect.setPlaceholder("No open role petitions");
+	}
 
-			for (const id in clubDictionary) {
-				const club = clubDictionary[id];
-				description += `\n__**${club.title}**__ (${club.userIds.length}${club.seats !== -1 ? `/${club.seats}` : ""} Members)\n**ID**: ${club.id}\n**Host**: <@${club.hostId}>\n`;
-				if (club.system) {
-					description += `**Game**: ${club.system}\n`;
+	[channelSelect, roleSelect].forEach(select => {
+		if (select.options.length > 0) {
+			select.setMaxValues(Math.min(select.options.length, MAX_EMBEDS_PER_MESSAGE));
+		} else {
+			select.setDisabled(true)
+				.addOptions([{
+					label: "no entries",
+					value: "no entries"
+				}])
+				.setMaxValues(1);
+		}
+	})
+
+	messageOptions.components = [new ActionRowBuilder().addComponents(channelSelect), new ActionRowBuilder().addComponents(roleSelect)];
+
+	const title = `Open Petition List (${commandMention("list petitions")})`;
+	const description = `Here are the open petitions for channels and pingable roles. They will automatically be added when reaching **${Math.ceil(memberCount * 0.05)} petitions** (5% of the server). You can sign an open petition with the select menus under this message.\n\nChannel petitions will be converted to lowercase because Discord text channels are all lowercase.`;
+	if (title.length + description.length + fields.reduce((total, field) => total + field.name.length + field.value.length, 0) > MAX_EMBED_TOTAL_CHARACTERS || fields.some(field => field.name.length > MAX_EMBED_FIELD_NAME_LENGTH || field.value.length > MAX_EMBED_FIELD_VALUE_LENGTH)) {
+		return new Promise((resolve, reject) => {
+			const fileText = [title, description, ...fields.map(field => `${field.name}\n${field.value}`)].join("\n\n");
+			fs.writeFile("data/listMessage.txt", fileText, "utf8", error => {
+				if (error) {
+					console.error(error);
 				}
-				if (club.timeslot.nextMeeting) {
-					description += `**Next Meeting**: <t:${club.timeslot.nextMeeting}>${club.timeslot.periodCount === 0 ? "" : ` repeats every ${club.timeslot.periodCount} ${club.timeslot.periodUnits === "weeks" ? "week(s)" : "day(s)"}`}\n`;
-				}
-				if (club.isRecruiting()) {
-					selectMenu.addOptions([
-						{
-							label: club.title,
-							description: `${club.userIds.length}${club.seats !== -1 ? `/${club.seats}` : ""} Members`,
-							value: club.id
-						}
-					])
-				}
-			}
+			});
+			resolve(messageOptions);
+		}).then(messageOptions => {
+			messageOptions.files = [{
+				attachment: "data/listMessage.txt",
+				name: "petitionList.txt"
+			}];
+			return messageOptions;
+		})
+	} else {
+		return new Promise((resolve, reject) => {
+			messageOptions.embeds = [
+				embedTemplateBuilder("#f07581")
+					.setTitle(title)
+					.setDescription(description)
+					.addFields(fields)
+			];
+			resolve(messageOptions);
+		})
+	}
+}
 
-			if (selectMenu.options.length > 0) {
-				selectMenu.setPlaceholder("Get club details...");
-			} else {
-				selectMenu.setPlaceholder("No clubs currently recruiting");
-			}
-			break;
+/** Builds the MessageOptions for the the club list message
+ * @returns {Promise<import('discord.js').BaseMessageOptions>}
+ */
+function buildClubListPayload() {
+	let description = "Here's a list of the clubs on the server. Learn more about one by using `/club-invite (club ID)`.\n";
+
+	const messageOptions = { flags: MessageFlags.SuppressNotifications };
+
+	const selectMenu = new StringSelectMenuBuilder().setCustomId("clubList")
+		.setMinValues(1);
+
+	for (const id in clubDictionary) {
+		const club = clubDictionary[id];
+		description += `\n__**${club.title}**__ (${club.userIds.length}${club.seats !== -1 ? `/${club.seats}` : ""} Members)\n**Host**: <@${club.hostId}>\n`;
+		if (club.system) {
+			description += `**Game**: ${club.system}\n`;
+		}
+		if (club.timeslot.nextMeeting) {
+			description += `**Next Meeting**: <t:${club.timeslot.nextMeeting}>${club.timeslot.periodCount === 0 ? "" : ` repeats every ${club.timeslot.periodCount} ${club.timeslot.periodUnits === "weeks" ? "week(s)" : "day(s)"}`}\n`;
+		}
+		if (club.isRecruiting()) {
+			selectMenu.addOptions([
+				{
+					label: club.title,
+					description: `${club.userIds.length}${club.seats !== -1 ? `/${club.seats}` : ""} Members`,
+					value: club.id
+				}
+			])
+		}
 	}
 
 	if (selectMenu.options.length > 0) {
-		// Max 10 because club summaries are shown as embeds and only 10 embeds fit in a single message
-		selectMenu.setMaxValues(Math.min(selectMenu.options.length, 10));
+		selectMenu.setPlaceholder("Get club details...");
+	} else {
+		selectMenu.setPlaceholder("No clubs currently recruiting");
+	}
+
+	if (selectMenu.options.length > 0) {
+		selectMenu.setMaxValues(Math.min(selectMenu.options.length, MAX_EMBEDS_PER_MESSAGE));
 	} else {
 		selectMenu.setDisabled(true)
 			.addOptions([{
@@ -226,9 +178,9 @@ function buildListMessagePayload(memberCount, listType) {
 			.setMaxValues(1);
 	}
 
-	messageOptions.components[0].addComponents(selectMenu);
+	messageOptions.components = [new ActionRowBuilder().addComponents(selectMenu)];
 
-	if (description.length > 2048) {
+	if (description.length > MAX_EMBED_DESCRIPTION_LENGTH) {
 		return new Promise((resolve, reject) => {
 			let fileText = description;
 			fs.writeFile("data/listMessage.txt", fileText, "utf8", error => {
@@ -240,7 +192,7 @@ function buildListMessagePayload(memberCount, listType) {
 		}).then(messageOptions => {
 			messageOptions.files = [{
 				attachment: "data/listMessage.txt",
-				name: `${listType}List.txt`
+				name: "clubList.txt"
 			}];
 			messageOptions.embeds = [];
 			return messageOptions;
@@ -249,10 +201,9 @@ function buildListMessagePayload(memberCount, listType) {
 		return new Promise((resolve, reject) => {
 			messageOptions.embeds = [
 				embedTemplateBuilder("#f07581")
-					.setTitle(listType === "club" ? `Clubs List (${commandMention("list clubs")})` : `Open Petitions List (${commandMention("list petitions")})`)
+					.setTitle(`Club List (${commandMention("list clubs")})`)
 					.setDescription(description)
 			];
-			messageOptions.files = [];
 			resolve(messageOptions);
 		})
 	}
@@ -263,7 +214,7 @@ function buildListMessagePayload(memberCount, listType) {
  * @param {"petition" | "club"} listType
  * @returns {Promise<Message>}
  */
-async function updateList(channelManager, listType) {
+async function updateListReference(channelManager, listType) {
 	const { channelId, messageId } = referenceMessages[listType];
 	if (channelId && messageId) {
 		const channel = await channelManager.fetch(channelId).catch(error => {
@@ -282,28 +233,18 @@ async function updateList(channelManager, listType) {
 			}
 			console.error(error);
 		});
-		const messageOptions = await buildListMessagePayload(channelManager.guild.memberCount, listType);
+		const messageOptions = listType === "club" ? await buildClubListPayload() : await buildPetitionListPayload(channelManager.guild.memberCount);
 		message?.edit(messageOptions);
-		if (messageOptions.files.length === 0) {
-			message?.removeAttachments();
-		}
 		return message;
 	}
 }
 
 module.exports = {
-	getPetitions,
-	setPetitions,
-	checkPetition,
-	getTopicIds,
-	getTopicNames,
-	addTopic,
-	removeTopic,
-	addTopicChannel,
 	getClubDictionary,
 	updateClub,
 	removeClub,
 	referenceMessages,
-	buildListMessagePayload,
-	updateList
+	buildPetitionListPayload,
+	buildClubListPayload,
+	updateListReference
 };