+++ ChangeLog.md
@@ -1,8 +1,8 @@
 # HorizonsBot Change Log
 #### HorizonsBot Version 2.8.0:
-- Proxy threads send silenced everyone mentions (then delete them) to simulate DM red bubbles
 - `/club-invite` now sends a message with a channel select menu and a user select menu instead of making users copy/paste the club's text channel id
 - Fixed `/at-channel` not creating notifications
+- `/petition pingable-role` can now be used to petition for pingable roles, old functionality moved to `/petition text-channel`
 
 #### HorizonsBot Version 2.7.0:
 ## /proxy-thread